preprocess_observation moved into its own file
