+++ pkmn.py
@@ -2,6 +2,7 @@
 import numpy as np
 import pickle    # I don't see any particular reason to remove pickle instead of writing to file some other way
 from env_pkmn import Env as pkmn_env
+from preprocess_observation import preprocess_observation as preprocess_observation
 
 # hyperparameters
 n = 14 # dimensionality of input 
@@ -58,78 +59,7 @@ def discount_rewards(r):
         discounted_r[t] = running_add
     return discounted_r
 
-"""Functions like the following four generally exist, but they are different based on game/model"""
-
-def preprocess_observation(I):
-    # There will always be a preprocessing step, but it will look different for different games.
-    # In this case, the string we get back from the Pokemon simulator does not give us the entire state
-    # of the game. Instead it gives us the change in the state. So return a list.
-    # Each element of the list contains two elements, the index of the state to update
-    # and the value to update it to.
-    retval = []
-    I = I.split('\n')
-    dbi = -1
-    ci = 0
-    retval2 = [-1,-1,-1,-1,-1,-1]
-    for line in I:
-        if ('switch|' in line) or ('drag|' in line):
-            # There is a new Pokemon on the field. Update the pokemon on field and the health.
-            if 'p1a' in line:
-                temp = line.split('|')
-                # Todo in maybe December: Make this handle all the Pokemon and not just our favorites
-                tempDict = {'Houndoom': 0, 'Ledian': 1, 'Lugia': 2, 'Malamar': 3, 'Swellow': 4, 'Victreebel': 5}
-                name = temp[2][5:]
-                index = tempDict[name]
-                retval.append([0, index])
-                health = int(temp[-1].split('/')[0])
-                retval.append([2 + index, health])
-            else:
-                assert('p2a' in line)
-                temp = line.split('|')
-                tempDict = {'Aggron': 0, 'Arceus': 1, 'Cacturne': 2, 'Dragonite': 3, 'Druddigon': 4, 'Uxie': 5}
-                name = temp[2][5:]
-                index = tempDict[name]
-                retval.append([1, index])
-                health = int(temp[-1].split('/')[0])
-                retval.append([8 + index, health])
-        elif 'damage' in line:
-            if 'Substitute' not in line:
-                tempDict = {'Houndoom': 2, 'Ledian': 3, 'Lugia': 4, 'Malamar': 5, 'Swellow': 6, 'Victreebel': 7, 'Aggron': 8, 'Arceus': 9, 'Cacturne': 10, 'Dragonite': 11, 'Druddigon': 12, 'Uxie': 13}
-                temp = line.split('|')
-                name = temp[2][5:]
-                if temp[-1][0] == '[':
-                    #The simulator is telling us the source of the damage
-                    health = 0
-                    if 'fnt' not in temp[-2]:
-                        health = int(temp[-2].split('/')[0])
-                    retval.append([tempDict[name], health])
-                else:
-                    if 'fnt' in temp[-1]:
-                        health = 0
-                        retval.append([tempDict[name], health])
-                    else:
-                        health = int(temp[-1].split('/')[0])
-                        retval.append([tempDict[name], health])
-        elif 'DEADBEEF' in line:
-            dbi = ci
-        elif line == 'p2: Aggron\r':
-            retval2[0] = ci
-        elif line == 'p2: Arceus\r':
-            retval2[1] = ci
-        elif line == 'p2: Cacturne\r':
-            retval2[2] = ci
-        elif line == 'p2: Dragonite\r':
-            retval2[3] = ci
-        elif line == 'p2: Druddigon\r':
-            retval2[4] = ci
-        elif line == 'p2: Uxie\r':
-            retval2[5] = ci
-        ci += 1
-
-    #There are way, way more parameters we can and should extract from this, but that's what we are doing for now
-    retval2 -= np.min(retval2)
-    retval2 += 1
-    return retval, retval2
+"""Functions like the following three generally exist, but they are different based on game/model"""
 
 def policy_forward(x):
     # Neural network begins here
@@ -273,6 +203,8 @@ def choose_action(x, bookkeeper):
         action = 'switch ' + str(bookkeeper.switch_indices[official_index])
     # Report to the bookkeeper the alphabetical index, but return the game index
     bookkeeper.report(x, h, pvec, action_index)
+    print(pvec)
+    print(action)
     return action
 
 def run_reinforcement_learning():
@@ -284,6 +216,7 @@ def run_reinforcement_learning():
         x = report_observation(observation)    
         action = choose_action(x, bookkeeper) 
         observation, reward, done, info = env.step(action)
+        #observation, reward, done, info = env.step("32")
         bookkeeper.report_reward(reward)
         if done: # an episode finished
             # Give backprop everything it could conceivably need
